KAS-2308 Delete unused code + moved separate renderers to util
